% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{adhesins_df}
\alias{adhesins_df}
\title{Adhesins gene-system data frame}
\format{
a data frame with two columns and 435 rows
}
\description{
Adhesins gene-system data frame
}
\keyword{datasets}
