% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{adhesins_df_grouped}
\alias{adhesins_df_grouped}
\title{Adhesins gene-system data frame with gene groups}
\format{
a data frame with two columns and 559 rows
}
\description{
Adhesins grouped gene-system data frame
}
\keyword{datasets}
