% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{adhesins_lengths}
\alias{adhesins_lengths}
\title{Adhesins gene length data frame}
\format{
a data frame with two columns and 525 rows
}
\description{
A data frame containing sequence length of each gene
}
\keyword{datasets}
