% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adhesiomeR-package.R
\docType{package}
\name{adhesiomeR-package}
\alias{adhesiomeR-package}
\alias{adhesiomeR}
\title{Analysis of *Escherichia coli* adhesiome}
\description{
For adhesiome analysis, adhesiomeR uses standalone BLAST and a manually curated 
database of adhesins. Our database contains **525 genes** comprising **102 systems** 
of fimbriae and other adhesins; therefore adhesiomeR provides the most 
comprehensive collection of *Escherichia coli* adhesins currently available.
}
\details{
AdhesiomeR is available as an R package and shiny GUI.
}
\author{
Maintainer: Katarzyna Sidorczuk <sidorczuk.katarzyna17@gmail.com>
}
\keyword{package}
