% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adhesiomeR_gui.R
\name{adhesiomeR_gui}
\alias{adhesiomeR_gui}
\title{Run adhesiomeR GUI}
\usage{
adhesiomeR_gui()
}
\value{
No return value, called for its side effects.
}
\description{
Launches a graphical user interface to adhesiomeR in a form of
a shiny application. The application may be used to perform adhesiomeR
analysis on up to 100 genomes.
}
