% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bitscore_thresholds}
\alias{bitscore_thresholds}
\title{Bitscore thresholds used in the strict mode}
\format{
a data frame with two columns and 525 rows
}
\description{
A data frame containing bitscore threshold for each gene
}
\keyword{datasets}
