% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{clustering_all}
\alias{clustering_all}
\title{Clustering based on all adhesins}
\format{
object of type clara, partition
}
\description{
Clustering of adhesin profiles based on all adhesins
}
\keyword{datasets}
