% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{clustering_fimbrial}
\alias{clustering_fimbrial}
\title{Clustering based on fimbrial adhesins}
\format{
object of type clara, partition
}
\description{
Clustering of adhesin profiles based on fimbrial adhesins
}
\keyword{datasets}
