% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{clustering_nonfimbrial}
\alias{clustering_nonfimbrial}
\title{Clustering based on nonfimbrial adhesins}
\format{
object of type clara, partition
}
\description{
Clustering of adhesin profiles based on nonfimbrial adhesins
}
\keyword{datasets}
