% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{df_genes}
\alias{df_genes}
\title{adhesiomeR genes information}
\format{
a data frame
}
\description{
A data frame with information about all genes in adhesiomeR
database, including system to which it belongs, sequence accession and 
function/class with references.
}
\keyword{datasets}
