% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{df_systems}
\alias{df_systems}
\title{adhesiomeR system information}
\format{
a data frame
}
\description{
A data frame with information about all systems in adhesiomeR
database, including its alternative names, type (fimbrial or nonfimbrial),
morphology (for fimbrial adhesins), biosynthesis/secretion pathway and
short description with references.
}
\keyword{datasets}
