% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blast_res.R
\name{do_blast_single}
\alias{do_blast_single}
\title{Get BLAST results for a single file}
\usage{
do_blast_single(input_file, blast_dir = Sys.which("blastn"))
}
\arguments{
\item{input_file}{Name of a file that will be used as input files for BLAST search. 
File have to contain nucleotide sequences in a FASTA format.}

\item{blast_dir}{A path to the directory with BLAST executables. By default,
it tries to find the proper executable by running system command \code{which}
(except on Windows). If the \code{blastn} executable is not found by default,
you should provide a path to the directory in which it is located.}
}
\value{
A data frame with BLAST results containing following information: 
query name, subject name, % of identity, alignment length, mismatches,
gap opens, query start, query end, subject start, subject end, E-value,
bit score (for detailed information about those outputs, please see BLAST
documentation), as well as name of a file that was used as an input.
}
\description{
This function runs BLAST search on a provided file.
}
\seealso{
get_blast_res
}
