% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomes.R
\name{gc_to_sample}
\alias{gc_to_sample}
\title{Expand gene catalogue results to samples}
\usage{
gc_to_sample(blast_res, abundance_matrix, threshold = NULL)
}
\arguments{
\item{blast_res}{a data frame with BLAST search results obtained by using
\code{\link{get_blast_res}}.}

\item{abundance_matrix}{A path to the abundance matrix file in
a tab-delimited format. The first column is expected to be named 'Gene' 
and contain gene names}

\item{threshold}{abundance threshold to consider a gene as present in a sample.
The default value is \code{NULL} and means that samples with any abundance
larger than 0 will be used.}
}
\value{
A data frame with BLAST results expanded to individual samples
}
\description{
This function extends blast results run on a gene catalogue to individual
samples that were used for gene catalogue creation.
}
\seealso{
get_blast_res
}
