% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Generate a HTML report with the results}
\usage{
generate_report(
  presence_table,
  elements = c("summary_table", "summary_plot", "presence_table", "presence_plot"),
  outdir = NULL,
  remove_intermediate_files = FALSE,
  hide_absent_genes = FALSE,
  hide_absent_systems = FALSE,
  presence_col = "#e42b24",
  absence_col = "#85c1ff"
)
}
\arguments{
\item{presence_table}{an object obtained by \code{\link{get_presence_table}}
function.}

\item{elements}{a character vector of elements which should be included
in the report. The available values are: \code{summary_table}, 
\code{summary_plot}, \code{presence_table} and \code{presence_plot}.
By default, all elements are used.}

\item{outdir}{a string indicating the path and name of the directory,
in which the report will be generated.}

\item{remove_intermediate_files}{a logical indicating if the files
with tables and/or plots used for report generation should be removed
from the results directory. By default \code{FALSE}.}

\item{hide_absent_genes}{a logical indicating if the genes that were
not found in any of the files should be hidden from the results.}

\item{hide_absent_systems}{a logical indicating if the systems that were
not found in any of the files should be hidden from the results.}

\item{presence_col}{color of the tiles representing present genes. Must be
specified as a hex color code}

\item{absence_col}{color of the tiles representing absent genes. Must be
specified as a hex color code}
}
\value{
No return value, called for its side effects.
}
\description{
This function generates a HTML report with the results of a performed
analysis, using an object obtained by \code{\link{get_presence_table}}
function.
}
\details{
This function uses object obtained by \code{get_presence_table}
function to create files with the results, i.e., plots and/or tables, 
depending on the selected elements. These files are placed in a newly 
created directory together with the report file, which location and name
may be specified using \code{outdir} argument. To get only the report file
in a results directory, \code{remove_intermediate_files} argument may be
used. 

The \code{elements} available to include in a report are:
\itemize{
  \item{summary_table}{ table with percents of genes found in each system}
  \item{summary_plot}{ plot with percents of genes found in each system}
  \item{presence_table}{ table with gene presence/absence}
  \item{presence_plot}{ plot with gene presence/absence.}}
}
