% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_adhesin_clusters}
\alias{get_adhesin_clusters}
\title{Assign genomes to adhesin profile clusters}
\usage{
get_adhesin_clusters(new_results, model_all, model_fimbrial, model_nonfimbrial)
}
\arguments{
\item{new_results}{a data frame of gene presence/absence generated using 
\code{\link{get_presence_table_strict}}}

\item{model_all}{clustering based on all adhesins}

\item{model_fimbrial}{clustering based on fimbrial adhesins}

\item{model_nonfimbrial}{clustering based on nonfimbrial adhesins}
}
\value{
a data frame with four columns containing file name and assignments
to clusters for three types of clustering (all adhesins, fimbrial, nonfimbrial).
}
\description{
Assigns analysed genomes to adhesin profile clusters using three subsets of genes:
all adhesins, fimbrial adhesins, nonfimbrial adhesins.
}
