% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blast_res.R
\name{get_blast_res}
\alias{get_blast_res}
\title{Get BLAST results for many files}
\usage{
get_blast_res(
  input_file_list,
  n_threads = 1,
  tmp_dir = "./",
  blast_dir = Sys.which("blastn")
)
}
\arguments{
\item{input_file_list}{A \code{list} of file names that will be used
as input files for BLAST search. File have to contain nucleotide sequences
in a FASTA format.}

\item{n_threads}{number of threads used for running BLAST. Default is one. The 
maximum number of threads is determined by \code{\link[parallel]{detectCores}}.}

\item{tmp_dir}{path to directory where temporary files will be stored while
running BLAST. By default current working directory.  It should be a directory
in which you have write access.}

\item{blast_dir}{A path to the directory with BLAST executables. By default,
it tries to find the proper executable by running system command \code{which}
(except on Windows). If the \code{blastn} executable is not found by default,
you should provide a path to the directory in which it is located.}
}
\value{
A data frame with BLAST results containing following information: 
query name, subject name, % of identity, alignment length, mismatches,
gap opens, query start, query end, subject start, subject end, E-value,
bit score (for detailed information about those outputs, please see BLAST
documentation), as well as name of a file that was used as an input.
}
\description{
This function runs BLAST search on a provided list of files.
}
\seealso{
do_blast_single
}
