% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes_systems_df.R
\name{get_genes_in_systems_db}
\alias{get_genes_in_systems_db}
\title{Get gene-system data frame}
\usage{
get_genes_in_systems_db(seq_file)
}
\arguments{
\item{seq_file}{File with sequences used for database creation 
in a FASTA format. See details section for FASTA header format requirements.}
}
\value{
a data frame containing systems and corresponding genes.
It consists of three columns:
\describe{
  \item{Gene}{Gene name}
  \item{System}{System to which given gene is assigned}}
Each row contains only one gene and system, so in case of genes assigned to 
multiple systems, they will occupy several rows.
}
\description{
Extract systems and corresponding genes from a sequence
file used for database creation.
}
\details{
The sequence file should be in a FASTA format with headers 
containing database name, gene name, accession, system name and each element
should be separated by \code{~} without spaces, e.g. 
\code{>Adhesins~fimH~NC_000913.3:4548808-4549710~Type_1_Fimbriae}. If a gene
is assigned to multiple systems, system names should be separated using 
semicolons without spaces, e.g. \code{Adhesins~aggR~NC_019000.1:48472-49269~AA/I_Fimbriae;
AA/II_Fimbriae;AA/III_Fimbriae}.
}
\examples{
get_genes_in_systems_db(system.file("./data/Adhesins_sequences.txt", package = "adhesiomeR"))

}
