% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_presence.R
\name{get_presence_plot}
\alias{get_presence_plot}
\title{Get plot with gene presence/absence}
\usage{
get_presence_plot(
  presence_table,
  systems = unique(adhesins_df[["System"]]),
  presence_col = "#e42b24",
  absence_col = "#85c1ff"
)
}
\arguments{
\item{presence_table}{a data frame with gene presence/absence obtained 
using \code{\link{get_presence_table}} function}

\item{systems}{a character vector with names of the systems that should
be shown on a heatmap. By default, all systems present in the database}

\item{presence_col}{color of the tiles representing present genes. Must be
specified as a hex color code}

\item{absence_col}{color of the tiles representing absent genes. Must be
specified as a hex color code}
}
\description{
This function generates a heatmap showing the presence or absence of each 
gene in each of the analyzed files.
}
