% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_presence.R
\name{get_presence_table}
\alias{get_presence_table}
\title{Get presence table}
\usage{
get_presence_table(
  blast_res,
  add_missing = TRUE,
  count_copies = FALSE,
  identity = 75,
  n_threads = 1
)
}
\arguments{
\item{blast_res}{blast results obtained with \code{\link{get_blast_res}}}

\item{add_missing}{\code{logical} indicating if genes not found by BLAST 
should be added. By default \code{TRUE}, meaning that all genes are shown 
in the resulting table, even if they were not found in any genome.}

\item{count_copies}{\code{logical} indicating if occurences of gene 
copies should be counted. An occurence of gene is considered a separate
copy if its location do not overlap with other hit to the same gene.}

\item{identity}{identity percent threshold}

\item{n_threads}{number of threads for parallel processing. Default is one. 
The maximum number of threads is determined by \code{\link[parallel]{detectCores}}.}
}
\value{
a data frame of gene presence/absence. The first column contains
the names of input files and the following correspond to analysed genes.
Presence of a gene is indicated by 1, whereas absence by 0. In case of
copy counts, numbers of gene copies is presented.
}
\description{
This function creates a presence/absence table of genes in analysed genomes
based on results of the BLAST search. Each row corresponds to one input file
and each column to one gene. Presence of a given gene is indicated as a 1, 
whereas absence as a 0. You can modify thresholds used to consider a gene
as present using \code{identity_threshold} and \code{evalue_threshold} 
arguments.
}
