% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_presence.R
\name{get_presence_table_relaxed}
\alias{get_presence_table_relaxed}
\title{Get presence table using relaxed settings}
\usage{
get_presence_table_relaxed(
  blast_res,
  add_missing = TRUE,
  count_copies = FALSE,
  identity = 75,
  coverage = 75,
  n_threads = 1
)
}
\arguments{
\item{blast_res}{blast results obtained with \code{\link{get_blast_res}}}

\item{add_missing}{\code{logical} indicating if genes not found by BLAST 
should be added. By default \code{TRUE}, meaning that all genes are shown 
in the resulting table, even if they were not found in any genome.}

\item{count_copies}{\code{logical} indicating if occurences of gene 
copies should be counted. An occurence of gene is considered a separate
copy if its location do not overlap with other hit to the same gene.}

\item{identity}{minimum identity percent threshold for considering gene
as present.}

\item{coverage}{minimum subject coverage per cent for considering gene
as present.}

\item{n_threads}{number of threads for parallel processing. Default is one. 
The maximum number of threads is determined by \code{\link[parallel]{detectCores}}.}
}
\value{
a data frame of gene presence/absence. The first column contains
the names of input files and the following correspond to analysed genes.
Presence of a gene is indicated by 1, whereas absence by 0. In case of
copy counts, numbers of gene copies is presented.
}
\description{
This function creates a presence/absence table using relaxed settings, i.e.
identity and coverage thresholds that may be modified by the user. This 
function performs initial filtering of the blast results according to set 
thresholds and calls \code{\link{get_presence_table}} to create the gene 
presence/absence table.
}
