% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary.R
\name{get_summary_plot}
\alias{get_summary_plot}
\title{Get plot with summarized system presence}
\usage{
get_summary_plot(
  presence_table,
  hide_absent = FALSE,
  presence_col = "#e42b24",
  absence_col = "#85c1ff"
)
}
\arguments{
\item{presence_table}{a data frame with gene presence/absence obtained 
using \code{\link{get_presence_table_strict}}
or \code{\link{get_presence_table_relaxed}} function}

\item{hide_absent}{\code{logical} indicating if columns representing 
systems that were not found in any file should be displayed. By default
\code{FALSE}}

\item{presence_col}{color of the tiles representing present genes. Must be
specified as a hex color code}

\item{absence_col}{color of the tiles representing absent genes. Must be
specified as a hex color code}
}
\description{
This function generates a heatmap showing the percent of genes found
for each system in each analysed file.
}
