% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_system_plot.R
\name{get_system_plot}
\alias{get_system_plot}
\title{Get plot with gene presence/absence for one system}
\usage{
get_system_plot(
  presence_table,
  system,
  presence_col = "#e42b24",
  absence_col = "#85c1ff"
)
}
\arguments{
\item{presence_table}{a data frame with gene presence/absence obtained 
using \code{\link{get_presence_table_strict}} or \code{\link{get_presence_table_relaxed}}
function}

\item{system}{\code{character} string indicating a name of selected system}

\item{presence_col}{color of the tiles representing present genes. Must be
specified as a hex color code}

\item{absence_col}{color of the tiles representing absent genes. Must be
specified as a hex color code}
}
\description{
This function creates a heatmap indicating gene presence/absence for one
selected system.
}
