% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{pathotype_colors}
\alias{pathotype_colors}
\title{Pathotype color palette}
\format{
An object of class \code{character} of length 13.
}
\usage{
pathotype_colors
}
\description{
Color palette for plots with pathotypes
}
\keyword{datasets}
