% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clustering.R
\name{plot_clustering_with_pathotypes}
\alias{plot_clustering_with_pathotypes}
\title{Plot clustering with pathotypes}
\usage{
plot_clustering_with_pathotypes(presence_table, show_labels = TRUE)
}
\arguments{
\item{presence_table}{data frame with gene presence/absence obtained 
using \code{\link{get_presence_table}} function.}

\item{show_labels}{logical indicating if labels with filenames of 
the analysed genomes should be plotted. Default is \code{TRUE}.}
}
\value{
A plot with the clustering results
}
\description{
This function clusters analysed genomes with pathotyped genomes
from RefSeq database. Clustering is performed using UMAP algorithm.
}
\seealso{
get_presence_table
}
