% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{profiles}
\alias{profiles}
\title{Adhesin profiles}
\format{
list of three data frames with profiles
}
\description{
Adhesin profiles based on all, fimbrial and nonfimbrial subsets
of adhesin genes
}
\keyword{datasets}
