% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomes.R
\name{gc_to_sample}
\alias{gc_to_sample}
\title{Expand gene catalogue results to samples}
\usage{
gc_to_sample(blast_res, abundance_matrix, threshold = NULL)
}
\arguments{
\item{blast_res}{a data frame with BLAST search results obtained by using
\code{\link{get_blast_res}}.}

\item{presence_absence_file}{A path to the abundance matrix file in
a tab-delimited format. The first column is expected to contain gene names.}
}
\value{
A data frame with BLAST results expanded to individual samples
}
\description{
This function extends blast results run on a gene catalogue to individual
samples that were used for gene catalogue creation.
}
\seealso{
get_blast_res
}
