% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_presence.R
\name{get_presence_table}
\alias{get_presence_table}
\title{Get presence table}
\usage{
get_presence_table(
  blast_res,
  add_missing = TRUE,
  identity_threshold = 70,
  evalue_threshold = 1e-50
)
}
\arguments{
\item{blast_res}{blast results obtained with \code{\link{get_blast_res}}}

\item{add_missing}{\code{logical} indicating if genes not found by BLAST 
should be added. By default \code{TRUE}, meaning that all genes are shown 
in the resulting table, even if they were not found in any genome.}

\item{identity_threshold}{\code{numeric} indicating the percent of identity
used for labeling a gene as present or absent}

\item{evalue_threshold}{\code{numeric} indicating the E-value threshold
used for labeling a gene as present or absent}
}
\value{
a data frame of gene presence/absence. The first column contains
the names of input files and the following correspond to analysed genes.
Presence of a gene is indicated by 1, whereas absence by 0.
}
\description{
This function creates a presence/absence table of genes in analysed genomes
based on results of the BLAST search. Each row corresponds to one input file
and each column to one gene. Presence of a given gene is indicated as a 1, 
whereas absence as a 0. You can modify thresholds used to consider a gene
as present using \code{identity_threshold} and \code{evalue_threshold} 
arguments. By default, gene is considered to be present when it shares
over 70% of identity with a subject sequence and has E-value lower than
1e-50.
}
