% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary.R
\name{get_summary_table}
\alias{get_summary_table}
\title{Get table with summarized system presence}
\usage{
get_summary_table(presence_table, hide_absent = FALSE)
}
\arguments{
\item{presence_table}{a data frame with gene presence/absence obtained 
using \code{\link{get_presence_table_strict}} or \code{\link{get_presence_table_relaxed}}
function}

\item{hide_absent}{\code{logical} indicating if columns representing 
systems that were not found in any file should be displayed. By default
\code{FALSE}}
}
\value{
a data frame with systems presence indicated by a percentage of
found genes. First column contains names of the input files and the following
correspond to analysed systems.
}
\description{
This functions creates a table with an overview of systems presence
in analysed genomes. Each system is summarized as a percent of genes
that were identified.
}
